+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Added database auditlog storage option [niravparikh05](https://github.com/niravparikh05)
+
 ## [0.1.7] - 2022-11-04
 
 ## Added